+++ aws-datastore/src/main/java/com/amplifyframework/datastore/syncengine/SyncProcessor.java
@@ -25,6 +25,7 @@ import com.amplifyframework.core.model.Model;
 import com.amplifyframework.core.model.ModelProvider;
 import com.amplifyframework.core.model.ModelSchema;
 import com.amplifyframework.core.model.ModelSchemaRegistry;
+import com.amplifyframework.datastore.AmplifyDisposables;
 import com.amplifyframework.datastore.DataStoreConfigurationProvider;
 import com.amplifyframework.datastore.DataStoreException;
 import com.amplifyframework.datastore.appsync.AppSync;
@@ -35,13 +36,11 @@ import com.amplifyframework.util.Time;
 import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicReference;
 
 import io.reactivex.Completable;
 import io.reactivex.Observable;
 import io.reactivex.Single;
 import io.reactivex.SingleEmitter;
-import io.reactivex.disposables.Disposable;
 import io.reactivex.schedulers.Schedulers;
 
 /**
@@ -167,41 +166,12 @@ final class SyncProcessor {
         return Single.<Iterable<ModelWithMetadata<T>>>create(emitter -> {
             final Cancelable cancelable =
                 appSync.sync(modelClass, lastSyncTimeAsLong, metadataEmitter(emitter), emitter::onError);
-            emitter.setDisposable(asDisposable(cancelable));
+            emitter.setDisposable(AmplifyDisposables.fromCancelable(cancelable));
         }).doOnSuccess(results ->
             LOG.debug("Successfully sync'd down cloud state for model type = " + modelClass.getSimpleName())
         ).doOnError(failureToSync ->
             LOG.warn("Failed to sync down cloud state for model type = " + modelClass.getSimpleName(), failureToSync)
         );
-    }
-
-    /**
-     * A utility method to convert a cancelable to a Disposable.
-     * TODO: Move this out to a more generic location?
-     * @param cancelable An Amplify Cancelable
-     * @return An RxJava2 Disposable that disposed by invoking the cancellation.
-     */
-    private static Disposable asDisposable(@NonNull Cancelable cancelable) {
-        Objects.requireNonNull(cancelable);
-        return new Disposable() {
-            private final AtomicReference<Boolean> isCanceled = new AtomicReference<>(false);
-            @Override
-            public void dispose() {
-                synchronized (isCanceled) {
-                    if (!isCanceled.get()) {
-                        cancelable.cancel();
-                        isCanceled.set(true);
-                    }
-                }
-            }
-
-            @Override
-            public boolean isDisposed() {
-                synchronized (isCanceled) {
-                    return isCanceled.get();
-                }
-            }
-        };
     }
 
     private static <T extends Model> Consumer<GraphQLResponse<Iterable<ModelWithMetadata<T>>>> metadataEmitter(
